+++ src/VBox/Main/src-client/RemoteUSBDeviceImpl.cpp
@@ -1,4 +1,4 @@
-/* $Id: RemoteUSBDeviceImpl.cpp 59381 2016-01-18 17:17:24Z alexander.eichner@oracle.com $ */
+/* $Id: RemoteUSBDeviceImpl.cpp 62370 2016-07-20 17:12:05Z noreply@oracle.com $ */
 
 /** @file
  *
@@ -78,7 +78,7 @@ HRESULT RemoteUSBDevice::init(uint32_t u32ClientId, VRDEUSBDEVICEDESC *pDevDesc,
     unconst(mData.backend)      = "vrdp";
 
     unconst(mData.port)         = pDevDesc->idPort;
-    unconst(mData.version)      = pDevDesc->bcdUSB >> 8;
+    unconst(mData.version)      = (uint16_t)(pDevDesc->bcdUSB >> 8);
     if (fDescExt)
     {
         VRDEUSBDEVICEDESCEXT *pDevDescExt = (VRDEUSBDEVICEDESCEXT *)pDevDesc;
@@ -116,7 +116,7 @@ HRESULT RemoteUSBDevice::init(uint32_t u32ClientId, VRDEUSBDEVICEDESC *pDevDesc,
     mData.state                  = USBDeviceState_Available;
 
     mData.dirty                  = false;
-    unconst(mData.devId)        = pDevDesc->id;
+    unconst(mData.devId)        = (uint16_t)pDevDesc->id;
 
     unconst(mData.clientId)     = u32ClientId;
